+++ resources/views/component/breadcrumb.blade.php
@@ -19,7 +19,19 @@
             @endif
         @endif
         @if(Request::segment(4))
-            <li class="breadcrumb-item"><a href="{{ url(Request::segment(1).'/'.Request::segment(2).'/'.Request::segment(3).'/'.Request::segment(4)) }}" class="text-decoration-none">{{ ucwords(Request::segment(4)) }}</a></li>
+            @if(!is_numeric(Request::segment(4)))
+                <li class="breadcrumb-item"><a href="{{ url(Request::segment(1).'/'.Request::segment(2).'/'.Request::segment(3).'/'.Request::segment(4)) }}" class="text-decoration-none">{{ ucwords(Request::segment(4)) }}</a></li>
+            @endif
+        @endif
+        @if(Request::segment(5))
+            @if(!is_numeric(Request::segment(5)))
+                <li class="breadcrumb-item"><a href="{{ url(Request::segment(1).'/'.Request::segment(2).'/'.Request::segment(3).'/'.Request::segment(4).'/'.Request::segment(5)) }}" class="text-decoration-none">{{ ucwords(Request::segment(5)) }}</a></li>
+            @endif
+        @endif
+        @if(Request::segment(6))
+            @if(!is_numeric(Request::segment(6)))
+                <li class="breadcrumb-item"><a href="{{ url(Request::segment(1).'/'.Request::segment(2).'/'.Request::segment(3).'/'.Request::segment(4).'/'.Request::segment(5).'/'.Request::segment(6)) }}" class="text-decoration-none">{{ ucwords(Request::segment(6)) }}</a></li>
+            @endif
         @endif
     </ol>
 </nav>